new db-transac

+++ internal/database/router.go
@@ -38,10 +38,9 @@ func (r *Router) AuthEndPoints() {
 }
 
 func (r *Router) UserEndPoints() {
-	r.R.HandleFunc("/", r.getList).Methods("GET")
-	r.R.HandleFunc("/{id}", r.getUserByID).Methods("GET")
-	r.R.HandleFunc("/{id}", r.updateUser).Methods("PUT")
-	r.R.HandleFunc("/{id}", r.partUpdateUser).Methods("PATCH")
+	r.R.HandleFunc("/{id}", r.UpdateAccount).Methods("PUT")
+	r.R.HandleFunc("/{id}", r.PartUpdateAccountName).Methods("PATCH")
+	r.R.HandleFunc("/{id}", r.PartUpdateAccountEmail).Methods("PATCH")
 	r.R.HandleFunc("/{id}", r.DeleteAccount).Methods("DELETE")
 }
 
@@ -53,26 +52,6 @@ func hello(w http.ResponseWriter, r *http.Request) {
 func (rout *Router) signIn(w http.ResponseWriter, r *http.Request) { // Entry
 	w.WriteHeader(http.StatusOK) // http_test.go
 	io.WriteString(w, "SignIn was successfully")
-}
-
-func (rout *Router) getList(w http.ResponseWriter, r *http.Request) { // GET
-	w.WriteHeader(http.StatusOK) // http_test.go
-	io.WriteString(w, "This is our users")
-}
-
-func (rout *Router) getUserByID(w http.ResponseWriter, r *http.Request) { // GET
-	w.WriteHeader(http.StatusOK) // http_test.go
-	io.WriteString(w, "This is our user by id")
-}
-
-func (rout *Router) updateUser(w http.ResponseWriter, r *http.Request) { // PUT
-	w.WriteHeader(204) // http_test.go
-	io.WriteString(w, "This is our updated user")
-}
-
-func (rout *Router) partUpdateUser(w http.ResponseWriter, r *http.Request) { // PATCH
-	w.WriteHeader(204) // http_test.go
-	io.WriteString(w, "This is our part update user")
 }
 
 func WriteJSON(w http.ResponseWriter, status int, a interface{}) error {
@@ -87,11 +66,13 @@ func (rout *Router) CreateAccount(w http.ResponseWriter, r *http.Request) {
 	if err := json.NewDecoder(r.Body).Decode(user); err != nil {
 		rout.logg.Errorln(err)
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
 	}
 
 	id, err := rout.rp.CreateNewUser(r.Context(), user)
 	if err != nil {
 		rout.logg.Errorln(err)
+		return
 	}
 
 	WriteJSON(w, http.StatusOK, map[string]interface{}{
@@ -100,15 +81,23 @@ func (rout *Router) CreateAccount(w http.ResponseWriter, r *http.Request) {
 }
 
 func (rout *Router) DeleteAccount(w http.ResponseWriter, r *http.Request) {
+	if r.Method != "DELETE" {
+		rout.logg.Errorln(http.StatusMethodNotAllowed)
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+
 	user := &User{}
 	if err := json.NewDecoder(r.Body).Decode(user); err != nil {
 		rout.logg.Errorln(err)
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
 	}
 
 	id, err := rout.rp.DeleteUser(r.Context(), user.ID)
 	if err != nil {
 		rout.logg.Errorln(err)
+		return
 	}
 
 	WriteJSON(w, http.StatusOK, map[string]interface{}{
@@ -116,14 +105,77 @@ func (rout *Router) DeleteAccount(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
-func GetByIdAccount(w http.ResponseWriter, r *http.Request) error {
-	return nil
+func (rout *Router) UpdateAccount(w http.ResponseWriter, r *http.Request) {
+	if r.Method != "PUT" {
+		rout.logg.Errorln(http.StatusMethodNotAllowed)
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+
+	user := &User{}
+	if err := json.NewDecoder(r.Body).Decode(user); err != nil {
+		rout.logg.Errorln(err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	id, err := rout.rp.UpdateUserFully(r.Context(), user)
+	if err != nil {
+		rout.logg.Errorln(err)
+		return
+	}
+
+	WriteJSON(w, http.StatusOK, map[string]interface{}{
+		"id": id,
+	})
 }
 
-func UpdateAccount(w http.ResponseWriter, r *http.Request) error {
-	return nil
+func (rout *Router) PartUpdateAccountName(w http.ResponseWriter, r *http.Request) {
+	if r.Method != "PATCH" {
+		rout.logg.Errorln(http.StatusMethodNotAllowed)
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+
+	user := &User{}
+	if err := json.NewDecoder(r.Body).Decode(user); err != nil {
+		rout.logg.Errorln(err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	id, err := rout.rp.PartUpdateUserName(r.Context(), user)
+	if err != nil {
+		rout.logg.Errorln(err)
+		return
+	}
+
+	WriteJSON(w, http.StatusOK, map[string]interface{}{
+		"id": id,
+	})
 }
 
-func PartUpdateAccount(w http.ResponseWriter, r *http.Request) error {
-	return nil
+func (rout *Router) PartUpdateAccountEmail(w http.ResponseWriter, r *http.Request) {
+	if r.Method != "PATCH" {
+		rout.logg.Errorln(http.StatusMethodNotAllowed)
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
+	}
+
+	user := &User{}
+	if err := json.NewDecoder(r.Body).Decode(user); err != nil {
+		rout.logg.Errorln(err)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	id, err := rout.rp.PartUpdateUserEmail(r.Context(), user)
+	if err != nil {
+		rout.logg.Errorln(err)
+		return
+	}
+
+	WriteJSON(w, http.StatusOK, map[string]interface{}{
+		"id": id,
+	})
 }
